requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
pandas>=2.2
numpy>=2.0
openpyxl>=3.1
requests>=2.31
gunicorn>=21.2.0
